requirements.txt
streamlit
watchdog
openai
boto3
ffmpeg-python
python-dotenv
# dev
ruff
aiohttp
